Code cleanup in pub_integration.
