+++ pkg/pub_integration/lib/src/fake_test_context_provider.dart
@@ -6,11 +6,13 @@ import 'dart:io';
 import 'dart:isolate';
 
 import 'package:_pub_shared/pubapi.dart';
+import 'package:http/http.dart' as http;
 import 'package:puppeteer/puppeteer.dart';
 
+import 'fake_credentials.dart';
 import 'fake_pub_server_process.dart';
-import 'fake_test_user.dart';
-import 'headless_env.dart';
+import 'pub_puppeteer_helpers.dart';
+import 'test_browser.dart';
 import 'test_scenario.dart';
 
 /// The timeout factor that should be used in integration tests.
@@ -65,11 +67,88 @@ class TestContextProvider {
     required String email,
     List<String>? scopes,
   }) async {
-    return await createFakeTestUser(
+    return await _createFakeTestUser(
       email: email,
       testBrowser: _testBrowser,
       fakeEmailReader: _fakePubServerProcess.fakeEmailReader,
       scopes: scopes,
     );
+  }
+}
+
+Future<TestUser> _createFakeTestUser({
+  required String email,
+  required TestBrowser testBrowser,
+  required FakeEmailReaderFromOutputDirectory fakeEmailReader,
+  List<String>? scopes,
+}) async {
+  late PubApiClient api;
+  await testBrowser.withPage(fn: (page) async {
+    await page.fakeAuthSignIn(email: email, scopes: scopes);
+    api = await _apiClientHttpHeadersFromSignedInSession(page);
+  });
+  return TestUser(
+    email: email,
+    api: api,
+    createCredentials: () => fakeCredentialsMap(email: email),
+    readLatestEmail: () async {
+      final map = await fakeEmailReader.readLatestEmail(recipient: email);
+      return map['bodyText'] as String;
+    },
+    withBrowserPage: <T>(Future<T> Function(Page) fn) async {
+      return await testBrowser.withPage<T>(fn: (page) async {
+        await page.fakeAuthSignIn(email: email, scopes: scopes);
+        return await fn(page);
+      });
+    },
+  );
+}
+
+/// Extracts the HTTP headers required for pub.dev API client (session cookies and CSRF token).
+Future<PubApiClient> _apiClientHttpHeadersFromSignedInSession(Page page) async {
+  await page.gotoOrigin('/my-liked-packages');
+  final csrfElement = await page.$('meta[name="csrf-token"]');
+  final csrfToken = await csrfElement.attributeValue('content');
+
+  final cookies = await page.cookies();
+  final cookieHeader = cookies
+      .where((e) => e.name.contains('-PUB_S') || e.name.startsWith('PUB_S'))
+      .map((e) => '${e.name}=${e.value}')
+      .join('; ');
+  final headers = <String, String>{
+    'cookie': cookieHeader,
+    if (csrfToken != null) 'x-pub-csrf-token': csrfToken,
+  };
+  final client = createHttpClientWithHeaders(headers);
+  return PubApiClient(page.origin, client: client);
+}
+
+http.Client createHttpClientWithHeaders(Map<String, String> headers) =>
+    _HttpClient(http.Client(), true, headers);
+
+/// An [http.Client] which sends additional headers along the request.
+class _HttpClient extends http.BaseClient {
+  final http.Client _client;
+  final bool _closeInnerClient;
+  final Map<String, String> _headers;
+
+  _HttpClient(
+    this._client,
+    this._closeInnerClient,
+    this._headers,
+  );
+
+  @override
+  Future<http.StreamedResponse> send(http.BaseRequest request) async {
+    request.headers.addAll(_headers);
+    return await _client.send(request);
+  }
+
+  @override
+  void close() {
+    if (_closeInnerClient) {
+      _client.close();
+    }
+    super.close();
   }
 }
