+++ pkg/pub_integration/lib/src/test_browser.dart
@@ -346,47 +346,3 @@ class _Coverage {
     return coveredPosition * 100 / textLength!;
   }
 }
-
-/// User to inject in the fake google auth JS script.
-class FakeGoogleUser {
-  final String? id;
-  final String? email;
-  final String? imageUrl;
-  final String? accessToken;
-  final String? idToken;
-  final String? scope;
-  final DateTime? expiresAt;
-
-  FakeGoogleUser({
-    this.id,
-    this.email,
-    this.imageUrl,
-    this.accessToken,
-    this.idToken,
-    this.scope,
-    this.expiresAt,
-  });
-
-  factory FakeGoogleUser.withDefaults(String email) {
-    final id = email.replaceAll('@', '-at-').replaceAll('.', '-dot-');
-    return FakeGoogleUser(
-      id: id,
-      email: email,
-      imageUrl: '/images/user/$id.jpg',
-      scope: 'profile',
-      accessToken: id,
-      idToken: id,
-      expiresAt: DateTime.now().add(Duration(hours: 1)),
-    );
-  }
-
-  Map<String, dynamic> toJson() => <String, dynamic>{
-        'isSignedIn': id != null,
-        'id': id,
-        'email': email,
-        'imageUrl': imageUrl,
-        'accessToken': accessToken,
-        'idToken': idToken,
-        'expiresAt': expiresAt?.millisecondsSinceEpoch ?? 0,
-      };
-}
